added unittesting for PriceAverage class
